fix: change getOS to function
